requirements.txt
numpy==2.2.2
opencv_python==4.11.0.86
Pillow==11.1.0
pyclipper==1.3.0.post6
PySide6==6.8.1.1
PySide6==6.8.2.1
PySide6_Addons==6.8.1.1
PySide6_Essentials==6.8.1.1
PyYAML==6.0.2
PyYAML==6.0.2
setuptools==75.8.0
Shapely==2.0.7
torch==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
